+++ mock-services/fake-service/build.sh
@@ -1,0 +1,2 @@
+sudo docker build . -t gordonwu/fake-service
+sudo docker push gordonwu/fake-service
