+++ mock-services/fake-service/package.json
@@ -1,0 +1,11 @@
+{
+  "scripts": {
+    "start": "node ratings.js"
+  },
+  "dependencies": {
+    "httpdispatcher": "1.0.0",
+    "faker": "4.1.0",
+    "mongodb": "^2.2.31",
+    "mysql": "^2.15.0"
+  }
+}
